Docs: Moved NavigationFooter to DocsPage
