+++ src/MudBlazor.Docs/Components/DocsPage.razor
@@ -4,6 +4,9 @@
             @ChildContent
         </CascadingValue>
     </div>
+    <div class="docs-navigation-footer">
+        <NavigationFooter Section=_section Previous="@_previous" Next="@_next" />
+    </div>
     <div class="docs-contents">
         <MudDrawer Open="true" Anchor="Anchor.End" Variant="DrawerVariant.Temporary" DisableOverlay="true" ClipMode="DrawerClipMode.Always" Elevation="0" Color="Color.Transparent" Class="pt-6" Width="220px">
             <MudPageContentNavigation SectionClassSelector="docs-section-header" @ref="_contentNavigation" ActivateFirstSectionAsDefault="true" />
