+++ src/MudBlazor.Docs/Components/NavigationFooter.razor
@@ -1,29 +1,27 @@
 ﻿@if (Section == NavigationSection.Components || Section == NavigationSection.Api)
  {
-    <MudContainer MaxWidth=@(Section==NavigationSection.Components?MaxWidth.Medium:MaxWidth.Large) Class="mb-0">
-            <MudDivider />
-            <MudToolBar DisableGutters="true">
-                @if (Previous!=null)
-                 {
-                    <MudButton Color="Color.Primary"
-                               Style="text-transform:none;"
-                               StartIcon="@Icons.Filled.NavigateBefore"
-                               Link="@($"/{Section.ToDescriptionString()}/{Previous.Link}")">
-                        @Previous.Name
-                    </MudButton>
-                 }
-                <MudSpacer/>
-                @if (Next!=null)
-                 {
-                    <MudButton Color="Color.Primary"
-                               Style="float:right;text-transform:none;"
-                               EndIcon="@Icons.Filled.NavigateNext"                          
-                               Link="@($"/{Section.ToDescriptionString()}/{Next.Link}")">
-                        @Next.Name
-                    </MudButton>
-                 }    
-            </MudToolBar>  
-    </MudContainer>
+    <MudDivider />
+    <MudToolBar DisableGutters="true">
+        @if (Previous!=null)
+            {
+            <MudButton Color="Color.Primary"
+                        Style="text-transform:none;"
+                        StartIcon="@Icons.Filled.NavigateBefore"
+                        Link="@($"/{Section.ToDescriptionString()}/{Previous.Link}")">
+                @Previous.Name
+            </MudButton>
+            }
+        <MudSpacer/>
+        @if (Next!=null)
+            {
+            <MudButton Color="Color.Primary"
+                        Style="float:right;text-transform:none;"
+                        EndIcon="@Icons.Filled.NavigateNext"                          
+                        Link="@($"/{Section.ToDescriptionString()}/{Next.Link}")">
+                @Next.Name
+            </MudButton>
+            }    
+    </MudToolBar>
  }
 
 @code{
