+++ src/MudBlazor.Docs/Shared/MainLayout.razor
@@ -88,7 +88,6 @@
                 <MudScrollToTop TopOffset="400" Style="z-index:2000;">
                     <MudFab Icon="@Icons.Material.Filled.KeyboardArrowUp" Color="Color.Primary" />
                 </MudScrollToTop>
-                <NavigationFooter Section=_section Previous="@_previous" Next="@_next" />
             </MudMainContent>
         </MudLayout>
     </MudSwipeArea>
